Extract cql executor interface (#2670)
* Instrument CqlExecutor

* [no release notes]
